docker/requirements.txt
kafka-python==1.4.5 
python-twitter